+++ tybotRoute/tiledeskChatbotPlugs/Filler.js
@@ -1,14 +1,19 @@
 var { Liquid } = require('liquidjs');
 var engine = new Liquid();
+const { v4: uuidv4 } = require('uuid');
 
 class Filler {
 
   fill(text, parameters) {
     // create dynamic attributes
-    if (parameters) {
-      parameters["timestamp"] = Date.now(); // type number
-      parameters["now"] = new Date(); // type Object  
+    if (!parameters) {
+      parameters = {};
     }
+
+    parameters["timestamp"] = Date.now(); // type number
+    parameters["now"] = new Date().toISOString(); // type Object
+    parameters["UUID"] = uuidv4().replace(/-/g, '');
+    parameters["UUIDv4"] = uuidv4();
 
     // legacy parser first
     if (text == null || text == undefined || typeof text !== 'string') {
@@ -22,13 +27,13 @@ class Filler {
     }
     
     // then post process with new LiquidJS!
-    let result;
+    let result = text;
     try {
       result = engine
       .parseAndRenderSync(text, parameters, null);
     }
     catch(e) {
-        console.error(e)
+      // console.error(e)
     }
     return result;
   }